Add some minor cosmetic changes
This commit may serve for future reference as to how some cosmetic
changes have been applied to the web.

Primarily using bootstrap classes in order to adjust margin or padding,
there are mechanisms that allow modifying the class of an
underlying html element in WTForms. Bootstrap classes are written
directly in html templates for Flask.

Also add a Soleta branded footer.