Add example image creation WTForms validator
Adds in a declarative way the validation of the image name form control
when creating a partition image.

This commit serves as an example of adding a predefined validator of the
WTForms module. Custom validator may be added by creating a given
function and appending it to the list of validators.

See https://wtforms.readthedocs.io/en/2.3.x/validators/#custom-validators