Add multiple ip handling to image restore form
This patch enables a set of ips to be specified as a target of an image
restore action.

The set of selected computers must have the same partition setup in
order for the form to be rendered sucessfully, if different partition
setups are detected a redirection is made and an error flash message
is shown.