Change login redirection to "Home"
When the user logs into ogcp, the web page redirects him to "Scopes".

In the future ogcp will have a dashboard showing multiple stats and
details. This dashboard will be in "Home" path and we want the user to
see it when he logs in.

Change login redirection from "Scopes" to "Home".