Create "Add Room" form and view
Creates "add room" form with required inputs only: center, name and
netmask.

Adds a new button inside the button group in the scopes view. Because
scopes checkboxes values maps to ips the only way to specify the center
in which to add the room is in the add room form itself, using a select
input.

In the future, the RoomForm can be used to display room information once
such functionality lands in the ogserver.