Add "Delete Room" form and view
Creates "delete room" form with a select to choose a room to delete.

Adds a new button inside the button group in the scopes view. Because
scopes checkboxes values maps to ips the only way to specify the room to
delete is in the delete room form itself, using a select input.