Fix setup form javascript
This patch fixes the javascript that adds/removes rows from the setup
form.

The AddPartition function now replaces the id `partitions-N-FIELDNAME`
using a regular expression. The hidden csrf input is also ignored to get
incremental numbers.

On deletion, all the ids are reassign from zero, in order, to get always
an ordered list with all numbers between the first and the last.