Add nav blocks to commands template
Adds empty jinja blocks for the class attribute of all action buttons.

Child templates from command (actions) can set their nav button as
active overriding the corresponding {% block %}.

Currently, not all action templates extends the commands.html template.
This is not a problem, as the blocks are just empty.

This is a preparation for changes on each action template to extend
commands.html template.